feat: remove mockData.ts e extrai tipos dedicados no CompanyDashboard
Move EsgPillar e InvestmentBudget para types.ts, atualiza os 4 imports
que apontavam para mockData, deleta mockData.ts e corrige index.test.tsx
para mockar fetchAuthenticatedProfile com companyName real.
